+++ src/main/main.c
@@ -454,12 +454,12 @@ static void sigactionSegv(int signum, siginfo_t *ip, void *context)
     /* First check for stack overflow if we know the stack position.
        We assume anything within 16Mb beyond the stack end is a stack overflow.
      */
-    if (signum == SIGSEGV && (ip != (siginfo_t *)0) && (long)R_CStackStart != -1)
+    if (signum == SIGSEGV && (ip != (siginfo_t *)0) && (intptr_t)R_CStackStart != -1)
     {
-        long addr = (long)ip->si_addr;
-        long diff = (R_CStackDir > 0) ? R_CStackStart - addr : addr - R_CStackStart;
-        long upper = 0x1000000; /* 16Mb */
-        if ((long)R_CStackLimit != -1)
+        uintptr_t addr = (uintptr_t)ip->si_addr;
+        intptr_t diff = (R_CStackDir > 0) ? R_CStackStart - addr : addr - R_CStackStart;
+        uintptr_t upper = 0x1000000; /* 16Mb */
+        if ((intptr_t)R_CStackLimit != -1)
             upper += R_CStackLimit;
         if (diff > 0 && diff < upper)
         {
